Bugfix in PDAFit:
Changed to a linear grid of epsilon values to ensure that the weights for individual epsilon contributions are the same. 
Before, a nonlinear grid in epsilon based on linear grid in E was used.

Since the returned P(eps) is a true probability density, the weights for the histogram contribution of a certain epsilon value is given by P(eps)deps! This was not done properly, since a nonlinear epsilon grid based on a linear grid in E was used. Changing to a linear grid of epsilon ensures that deps is the same for all epsilons, thus the probability density directly gives the desired weight.

git-svn-id: svn://10.153.116.2/PAM@1192 cb18dcf6-275f-e045-ae08-2e8c34a2df07